+++ netbox/dcim/exceptions.py
@@ -1,0 +1,5 @@
+class LoopDetected(Exception):
+    """
+    A loop has been detected while tracing a cable path.
+    """
+    pass