+++ CHANGELOG.md
@@ -14,6 +14,7 @@ v2.5.2 (FUTURE)
 * [#2707](https://github.com/digitalocean/netbox/issues/2707) - Correct permission evaluation for circuit termination cabling
 * [#2712](https://github.com/digitalocean/netbox/issues/2712) - Preserve list filtering after editing objects in bulk
 * [#2717](https://github.com/digitalocean/netbox/issues/2717) - Fix bulk deletion of tags
+* [#2721](https://github.com/digitalocean/netbox/issues/2721) - Detect loops when tracing front/rear ports
 * [#2723](https://github.com/digitalocean/netbox/issues/2723) - Correct permission evaluation when bulk deleting tags
 
 ---